Update README.md
This changes README.md from referring Firefox to referring to IceWeasel, and adds a little information about the new Iceweasel project.

For now, a lot of the technical details present in the old and upstream versions of this file are missing, because I didn't feel I had enough understanding of the concepts and technical details to edit them for relevance to this project.  My thought is that they could be revised later by someone who understands them better and merged back into this document as part of a future update to the document.  This is in some respects a placeholder and a starting point.