Hide private browsing myths link
Also turn off Android test for it.

This is easier than actually writing our own documentation, disentangling the
link from the SUMO topic linking system, and changing the tests to respect the
new destination. The downside is that if users don't know what they're doing
with private browsing, we will no longer even try to help them.

This fixes #221.